requirements.txt
Flask
Pillow
pytesseract
openpyxl
